Reorganise hostcalls into submodules
